removed distribute resources from hex class
